+++ src/tags/audio/opus.rs
@@ -1,7 +1,0 @@
-//! OPUS 音频格式标签
-
-use crate::TagId;
-
-pub const COMMENTS: TagId = TagId("Comments");
-pub const OUTPUT_GAIN: TagId = TagId("OutputGain");
-pub const SAMPLE_RATE: TagId = TagId("SampleRate");