+++ src/tags/formats/dng.rs
@@ -1,17 +1,12 @@
-//! DNG 图像格式标签
+#![allow(non_upper_case_globals)]
+
+//! 标签定义 - formats/dng
 
 use crate::TagId;
 
-pub const HINT_MAX_OUTPUT_VALUE: TagId = TagId("HintMaxOutputValue");
-pub const MAKER_NOTES: TagId = TagId("MakerNotes");
-pub const ORIGINAL_RAW_IMAGE: TagId = TagId("OriginalRawImage");
-pub const ORIGINAL_RAW_RESOURCE: TagId = TagId("OriginalRawResource");
-pub const ORIGINAL_T_H_M_IMAGE: TagId = TagId("OriginalTHMImage");
-pub const ORIGINAL_T_H_M_RESOURCE: TagId = TagId("OriginalTHMResource");
-pub const P_D_R_VERSION: TagId = TagId("PDRVersion");
-pub const SEQ_COUNT: TagId = TagId("SeqCount");
-pub const SEQ_FINAL: TagId = TagId("SeqFinal");
-pub const SEQ_FRAME_INFO: TagId = TagId("SeqFrameInfo");
-pub const SEQ_ID: TagId = TagId("SeqID");
-pub const SEQ_INDEX: TagId = TagId("SeqIndex");
-pub const SEQ_TYPE: TagId = TagId("SeqType");
+pub const DNGAdobeData: TagId = TagId("DNGAdobeData");
+pub const DNGBackwardVersion: TagId = TagId("DNGBackwardVersion");
+pub const DNGIgnoreSidecars: TagId = TagId("DNGIgnoreSidecars");
+pub const DNGLensInfo: TagId = TagId("DNGLensInfo");
+pub const DNGPrivateData: TagId = TagId("DNGPrivateData");
+pub const DNGVersion: TagId = TagId("DNGVersion");