+++ src/tags/formats/gif.rs
@@ -1,16 +1,11 @@
-//! GIF 图像格式标签
+#![allow(non_upper_case_globals)]
+
+//! 标签定义 - formats/gif
 
 use crate::TagId;
 
-pub const ANIMATION: TagId = TagId("Animation");
-pub const ANIMATION_ITERATIONS: TagId = TagId("AnimationIterations");
-pub const BITS_PER_PIXEL: TagId = TagId("BitsPerPixel");
-pub const CHANNEL_USAGE: TagId = TagId("ChannelUsage");
-pub const COLOR_RESOLUTION_DEPTH: TagId = TagId("ColorResolutionDepth");
-pub const DELAY_TIME: TagId = TagId("DelayTime");
-pub const HAS_COLOR_MAP: TagId = TagId("HasColorMap");
-pub const I_C_C_PROFILE: TagId = TagId("ICC_Profile");
-pub const J_U_M_B_F: TagId = TagId("JUMBF");
-pub const M_ID_I_SONG: TagId = TagId("MIDISong");
-pub const PIXEL_ASPECT_RATIO: TagId = TagId("PixelAspectRatio");
-pub const X_M_P: TagId = TagId("XMP");
+pub const GIFApplicationExtension: TagId = TagId("GIFApplicationExtension");
+pub const GIFGraphicControlExtension: TagId = TagId("GIFGraphicControlExtension");
+pub const GIFPlainTextExtension: TagId = TagId("GIFPlainTextExtension");
+pub const GIFTFtpPriority: TagId = TagId("GIFTFtpPriority");
+pub const GIFVersion: TagId = TagId("GIFVersion");