+++ src/tags/formats/ico.rs
@@ -1,9 +1,0 @@
-//! ICO 图像格式标签
-
-use crate::TagId;
-
-pub const BITS_PER_PIXEL: TagId = TagId("BitsPerPixel");
-pub const COLOR_PLANES: TagId = TagId("ColorPlanes");
-pub const HOTSPOT_X: TagId = TagId("HotspotX");
-pub const HOTSPOT_Y: TagId = TagId("HotspotY");
-pub const IMAGE_LENGTH: TagId = TagId("ImageLength");