+++ src/tags/standard/xmp.rs
@@ -1,21 +1,12 @@
-//! 标签定义
+#![allow(non_upper_case_globals)]
+
+//! 标签定义 - standard/xmp
 
 use crate::TagId;
 
-pub const XMP_DC_TITLE: TagId = TagId("Title");
-pub const XMP_DC_CREATOR: TagId = TagId("Creator");
-pub const XMP_DC_SUBJECT: TagId = TagId("Subject");
-pub const XMP_DC_DESCRIPTION: TagId = TagId("Description");
-pub const XMP_DC_PUBLISHER: TagId = TagId("Publisher");
-pub const XMP_DC_CONTRIBUTOR: TagId = TagId("Contributor");
-pub const XMP_DC_DATE: TagId = TagId("Date");
-pub const XMP_DC_TYPE: TagId = TagId("Type");
-pub const XMP_DC_FORMAT: TagId = TagId("Format");
-pub const XMP_DC_IDENTIFIER: TagId = TagId("Identifier");
-pub const XMP_DC_SOURCE: TagId = TagId("Source");
-pub const XMP_DC_LANGUAGE: TagId = TagId("Language");
-pub const XMP_DC_RELATION: TagId = TagId("Relation");
-pub const XMP_DC_COVERAGE: TagId = TagId("Coverage");
-pub const XMP_DC_RIGHTS: TagId = TagId("Rights");
-pub const XMP_XMP_RIGHTS_WEB_STATEMENT: TagId = TagId("WebStatement");
-pub const XMP_XMP_RIGHTS_USAGE_TERMS: TagId = TagId("UsageTerms");
+pub const EmbeddedXMPDigest: TagId = TagId("EmbeddedXMPDigest");
+pub const HasExtendedXMP: TagId = TagId("HasExtendedXMP");
+pub const LastKeywordXMP: TagId = TagId("LastKeywordXMP");
+pub const OldXMP: TagId = TagId("OldXMP");
+pub const XMP: TagId = TagId("XMP");
+pub const XMPToolkit: TagId = TagId("XMPToolkit");